+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- `impl SpanRanged for Range<impl SpanRanged>`
+
 ## [0.10.2] - 2023-11-20
 ### Added
 - `SpanRanged::span_joined` a function to return joined spans on nightly (to replace `SpanRanged::joined`).